pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]
